add last source

+++ migrations/sqls/20230514070818-products-table-up.sql
@@ -2,5 +2,5 @@ CREATE TABLE products (
     id SERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     description TEXT,
-    price DECIMAL(10, 2) NOT NULL
+    price INTEGER NOT NULL
 );
